forward/requirements.txt
Flask==1.1.2
paho_mqtt==1.6.1
